requirements.txt
Django==1.9.10
django-debug-toolbar==1.6
django-money==0.9.1
psycopg2==2.6.2
py-moneyed==0.6.0
sqlparse==0.2.1
